requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pymupdf>=1.23.0
pyyaml>=6.0
httpx>=0.25.0
